API/requirements.txt
bcrypt==3.1.7
boto==2.49.0
boto3==1.12.39
botocore==1.15.39
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
gensim==3.8.1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.5
MarkupSafe==1.1.1
multi-key-dict==2.0.3
nltk==3.4.5
numpy==1.18.2
pandas==1.0.3
pbr==5.4.2
pycparser==2.19
pymongo==3.9.0
python-dateutil==2.8.1
python-jenkins==1.5.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
s3transfer==0.3.3
scipy==1.4.1
six==1.12.0
smart-open==1.11.1
urllib3==1.25.3
Werkzeug==0.15.5
oauth2==1.9.0.post1
